Extract DSL modules from tracker
