+++ packages/data_tracker/lib/data_tracker.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "data_tracker/dsl"
 require_relative "data_tracker/version"
 require_relative "data_tracker/tracker"
 
@@ -13,21 +14,12 @@ module DataTracker
   end
 
   def self.setup(&block)
-    ::DataTracker::Builder.module_eval(&block)
+    ::DataTracker::DSL.build(&block)
   end
 
   def self.apply(model)
     @trackers.each do |_key, tracker|
       tracker.apply(model)
-    end
-  end
-
-  # :nodoc
-  module Builder
-  module_function
-
-    def tracker(key, &block)
-      DataTracker.trackers[key] = Tracker.new(&block)
     end
   end
 end
